fix: rename to shoppingcart to re-use the shoppingcart tables
