+++ README.md
@@ -1,10 +1,12 @@
 # appsembler_api
 
+NOTE: the app is called `shoppingcart` to hack in the old shoppingcart tables from pre-Juniper.
+
 Provide legacy (Hawthorn and earlier) appsembler_api as an LMS Django app plugin
 for use in Juniper and later (possibly).
 
 This API is deprecated though and ideally its functionality will be refactored to
-openedx.core.djangoapps.appsembler.api in our forks of edx-platform, or as moved to 
+openedx.core.djangoapps.appsembler.api in our forks of edx-platform, or as moved to
 another service.
 
 Currently maintained for standalone customers.
