+++ shoppingcart/__init__.py
@@ -7,4 +7,4 @@ Provides account, enrollment, analytics APIs.
 __version__ = '0.1.0'
 
 
-default_app_config = 'appsembler_api.apps.AppsemblerApiConfig'  # pylint: disable=invalid-name
+default_app_config = 'shoppingcart.apps.AppsemblerApiConfig'  # pylint: disable=invalid-name
