+++ shoppingcart/utils.py
