+++ tox.ini
@@ -32,28 +32,28 @@ match-dir = (?!migrations)
 
 [pytest]
 DJANGO_SETTINGS_MODULE = test_settings
-addopts = --cov appsembler_api --cov-report term-missing --cov-report xml
+addopts = --cov shoppingcart --cov-report term-missing --cov-report xml
 norecursedirs = .* docs requirements site-packages
 
 [testenv]
 basepython=python3
-deps = 
+deps =
    django2: Django>=2.2,<3
    -rrequirements/test.txt
-;commands = 
+;commands =
 ;    python manage.py check
 ;    pytest {posargs}
 
 [testenv:quality]
 basepython=python3
-whitelist_externals = 
+whitelist_externals =
     make
     rm
     touch
-deps = 
+deps =
     -rrequirements/quality.txt
-commands = 
+commands =
     touch tests/__init__.py
-    pylint --rcfile=pylintrc appsembler_api tests test_utils manage.py setup.py
+    pylint --rcfile=pylintrc shoppingcart tests test_utils manage.py setup.py
     rm tests/__init__.py
     make selfcheck
